+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* TheAlgorithm.reset() also resets concurrency
 
 ### v0.25.3
 * Add a `source` property to `Toot` class